Various fixes
ConveyorButtonControl.cpp: not sure what ":
CommandBase("ConveyorButtonControl")" does, but all the other commands
had it
ConveyorButtonControl.h: fix path to CommandBase.h, get rid of
"controlButtonID"
HopperWithGamepad.cpp: post hopper encoder value to SmartDashboard for
testing purposes
TeleopControl.cpp: post left and right encoder values to SmartDashboard
for testing purposes
definitions.h: the constants for the hopper encoder channels had been
incorrectly named as the conveyor encoders (and the conveyor does not
use an encoder)
Robot.cpp: add ConveyorButtonControl to the Scheduler; otherwise it
would not run. also some clarification in the comments
Drive.cpp: the left encoder had been using the right encoder channels.
change the GetLeftEncoderValue and GetRightEncoderValue functions from
int to double
Drive.h: change the GetLeftEncoderValue  and GetRightEncoderValue
functions from int to double
Hopper.cpp: the constants for the encoder channels had been incorrectly
named conveyor

+++ Team Komodo 2017/src/Subsystems/Drive.h
@@ -23,8 +23,8 @@ public:
 	void Stop();
 	void ResetLeftEncoder();
 	void ResetRightEncoder();
-	int GetLeftEncoderValue();
-	int GetRightEncoderValue();
+	double GetLeftEncoderValue();
+	double GetRightEncoderValue();
 	Encoder* GetLeftEncoder();
 	Encoder* GetRightEncoder();
 };
